[REFACTOR] TotalResultItem helper file
